multiple minor changes, optimizations

+++ app/src/main/java/tech/techlore/plexus/models/main/MainData.kt
@@ -40,7 +40,7 @@ data class MainData(
     
     var installedVersion: String = "",
     
-    var installedBuild: Int = 0,
+    var installedBuild: Long = 0L,
     
     var installedFrom: String = "",
