+++ app/src/main/java/tech/techlore/plexus/models/myratings/MyRatingDetails.kt
@@ -28,7 +28,7 @@ data class MyRatingDetails(
     
     var version: String,
     
-    var buildNumber: Int,
+    var buildNumber: Long,
     
     var romName: String,
